Add ConnectEntities and import PLAYER
Added a new controller for connecting entities and imported the PLAYER variable from the factorio_instance module to use it in the code. This controller allows the user to connect two entities on the basis of their locations, their connection type, and the whether their locations should be considered relative or absolute.