Fix electric wire connection distance in connect_entities.lua
In this commit, we fix a bug in connect_entities.lua where the electric wire connection distance did not match the correct value. We change how connector_length is calculated by setting it to 1 by default and then checking if the connector has wire connection points and setting it to the maximum wire distance in that case. We also remove some commented code used for debugging purposes.