Fix issue with creating beam points in connect_entities.lua (#44)
Removed unnecessary commented out code that was preventing the creation of beam points in the connect_entities.lua file. Additionally, fixed an offset bug that was causing incorrect placement of entities along a path in the same file. Added a check for step size to determine the initial offset for entity placement.